refactor: move GetTxClientFromFlags to deps/config pkg
