+++ x/migration/module/cmd/claim_morse_supplier_bulk.go
@@ -18,8 +18,9 @@ import (
 
 	"github.com/pokt-network/poktroll/cmd/flags"
 	"github.com/pokt-network/poktroll/cmd/logger"
+	"github.com/pokt-network/poktroll/pkg/deps/config"
 	"github.com/pokt-network/poktroll/x/migration/types"
-	"github.com/pokt-network/poktroll/x/supplier/config"
+	supplierconfig "github.com/pokt-network/poktroll/x/supplier/config"
 )
 
 var (
@@ -378,7 +379,7 @@ func runClaimSuppliers(cmd *cobra.Command, _ []string) error {
 	}
 
 	// Construct a tx client.
-	txClient, err := flags.GetTxClientFromFlags(ctx, cmd)
+	txClient, err := config.GetTxClientFromFlags(ctx, cmd)
 	if err != nil {
 		return err
 	}
@@ -477,7 +478,7 @@ func getMorseAccountsFromFile(morseNodesFile string) ([]MorseAccountInfo, error)
 // - Stake amount is omitted (set by protocol logic).
 // - Owner/operator addresses should not be set in the template.
 // - Only service configs and revenue share settings should be included.
-func loadTemplateSupplierStakeConfigYAML(configYAMLPath string) (*config.YAMLStakeConfig, error) {
+func loadTemplateSupplierStakeConfigYAML(configYAMLPath string) (*supplierconfig.YAMLStakeConfig, error) {
 	// Read the YAML file from the provided path.
 	yamlStakeConfigBz, err := os.ReadFile(configYAMLPath)
 	if err != nil {
@@ -485,7 +486,7 @@ func loadTemplateSupplierStakeConfigYAML(configYAMLPath string) (*config.YAMLSta
 	}
 
 	// Unmarshal the YAML into a config.YAMLStakeConfig struct.
-	var yamlStakeConfig config.YAMLStakeConfig
+	var yamlStakeConfig supplierconfig.YAMLStakeConfig
 	if err = yaml.Unmarshal(yamlStakeConfigBz, &yamlStakeConfig); err != nil {
 		return nil, err
 	}
@@ -555,8 +556,8 @@ func queryMorseClaimableAccount(
 func buildSupplierStakeConfig(
 	ownerAddress string,
 	operatorAddress string,
-	templateSupplierStakeConfig *config.YAMLStakeConfig,
-) (*config.SupplierStakeConfig, error) {
+	templateSupplierStakeConfig *supplierconfig.YAMLStakeConfig,
+) (*supplierconfig.SupplierStakeConfig, error) {
 	if ownerAddress == "" {
 		return nil, fmt.Errorf("owner address must be non-empty")
 	}
@@ -607,7 +608,7 @@ func buildSupplierStakeConfig(
 		return nil, err
 	}
 
-	return &config.SupplierStakeConfig{
+	return &supplierconfig.SupplierStakeConfig{
 		OwnerAddress:    yamlStakeConfig.OwnerAddress,
 		OperatorAddress: yamlStakeConfig.OperatorAddress,
 		Services:        supplierServiceConfigs,
